Alterando migration, criando factory e seeder

+++ database/migrations/2024_03_21_122903_create_funcionarios_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nome');
             $table->string('email');
-            $table->decimal('salario', 10, 2); // 10 dígitos no total, 2 após a vírgula
+            $table->float('salario');
             $table->timestamps();
         });
     }
